requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.1.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
livereload==2.7.0
MarkupSafe==3.0.2
packaging==24.2
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
tornado==6.4.2
typing_extensions==4.12.2
Werkzeug==3.1.3
